stdlib: enumerate produces i64 keys, matching its declared type
array::enumerate and list::enumerate built their key from the Rust
usize iterator index, which converts to Value::U64 — but both declare
(i64, %s) pairs. The node-walk doesn't enforce declared types so it
shipped U64 keys; a fused kernel computes the key natively AS the
declared i64 — a value-type divergence (interp [U64(1),U64(3),U64(5)]
vs jit [I64(...)] for map(enumerate(a), |(k,v)| k+v), soak jul08g
divergences 5 and 7). The declared type is authoritative; the builtin
was lying. Cast the index.

+++ stdlib/graphix-package-array/src/lib.rs
@@ -1089,7 +1089,7 @@ impl<R: Rt, E: UserEvent> EvalCached<R, E> for EnumerateEv {
     fn eval(&mut self, _ctx: &mut ExecCtx<R, E>, from: &CachedVals) -> Option<Value> {
         if let Some(Value::Array(a)) = &from.0[0] {
             let a = ValArray::from_iter_exact(
-                a.iter().enumerate().map(|(i, v)| (i, v.clone()).into()),
+                a.iter().enumerate().map(|(i, v)| (i as i64, v.clone()).into()),
             );
             return Some(Value::Array(a));
         }
